[Filters] Fix bug where changes to `appliedFilters` prop don't get respected (#10566)
### WHY are these changes introduced?

Navigating through the admin this morning I noticed an issue on the
Orders index, when refreshing the page with a set of filters that don't
match any existing filter, that the active filters don't appear in the
Filters component. See video below:



https://github.com/Shopify/polaris/assets/2562596/bb170b1f-0739-4778-8474-54c1ffe29f6e

It turns out that the Orders index async loads the filter data if it
does not match a saved view, and returns an empty `appliedFilters` array
initially, before updating it with the correct `appliedFilters`.

Due to [this recently introduced
change](https://github.com/Shopify/polaris/pull/10429) (sorry folks 🤡)
we no longer are able to react to changes to the `appliedFilters` prop
because of [this
line](https://github.com/Shopify/polaris/pull/10429/files#diff-f21b98ea8a1f295706a4342f1e9ce543562f6c47d6d0682367e90bb7c23961d8R167),
which sets the `localPinnedFilters` state only when the component
mounts. In the case of the Orders index with the parameters mentioned
above, this would set 0 pinned filters, due to the `appliedFilters`
being empty. Because we don't call `setLocalPinnedFilters` on changes to
the `appliedFilters`, this gives us the bug outlined in the video above.
The filters do get set, but the filter does not appear in the pinned
list.


### WHAT is this pull request doing?

This PR effectively reverts that change, and instead uses a more robust
approach to ensure that FilterPills being powered by `appliedFilters`
rather than `localPinnedState` do not vanish when they become unapplied
(when the filter values for the filter are removed).

What we want to do, is add any applied filters to the array of
`localPinnedState` so that when any filter value gets removed, the
FilterPill remains as an unselected FilterPill. The merchant can then
click out of the Popover to fully remove the FilterPill from the Filters
bar.

This is achieved by a `useEffect` hook, which listens for changes to the
`appliedFilterKeys` and `localPinnedFilters` arrays, and checks if all
values within `appliedFilterKeys` are not currently in the
`localPinnedFilters` array. If that is the case, we know that we want to
add these keys to the `localPinnedFilters` array, to ensure they are
added as `FilterPill`s if the `appliedFilters` value changes post-mount.

### How to 🎩

🖥 [Local development
instructions](https://github.com/Shopify/polaris/blob/main/README.md#local-development)
🗒 [General tophatting
guidelines](https://github.com/Shopify/polaris/blob/main/documentation/Tophatting.md)
📄 [Changelog
guidelines](https://github.com/Shopify/polaris/blob/main/.github/CONTRIBUTING.md#changelog)

Video of Snapshot in use on Spinstance:


https://github.com/Shopify/polaris/assets/2562596/23decb34-c423-412c-90f0-213d89834786

- Spinstance:
https://admin.web.orders-pill-check.marc-thomas.eu.spin.dev/store/shop1/orders?inContextTimeframe=none
- Note that adding in filters that don't match a saved view, and then
refreshing, the applied filters appear as FilterPills.

- Storybook:
https://5d559397bae39100201eedc1-kqkexazlai.chromatic.com/?path=/story/all-components-indexfilters--with-no-pinned-and-prefilled-filters&globals=polarisSummerEditions2023:true
- Note that removing the values from the "Account status" filter keeps
the filter pill visible, until the Popover is closed

### 🎩 checklist

- [x] Tested on
[mobile](https://github.com/Shopify/polaris/blob/main/documentation/Tophatting.md#cross-browser-testing)
- [x] Tested on [multiple
browsers](https://help.shopify.com/en/manual/shopify-admin/supported-browsers)
- [x] Tested for
[accessibility](https://github.com/Shopify/polaris/blob/main/documentation/Accessibility%20testing.md)
- [x] Updated the component's `README.md` with documentation changes
- [x] [Tophatted
documentation](https://github.com/Shopify/polaris/blob/main/documentation/Tophatting%20documentation.md)
changes in the style guide

---------

Co-authored-by: Chloe Rice <chloerice@users.noreply.github.com>