fix: Stop Filter hitting max call depth when appliedFilters is undefined
Currently `appliedFilters` can be `undefined`, except when it is
when we try to `setLocalPinnedFilters`, we'll do so forever because we
fail the check to decide whether to try set state or not every time, as
`!allAppliedFilterKeysInLocalPinnedFilters` will always be `true`.

To avoid this, we should only try to make the local pinned filters match
if we actually have any applied filters - and it's not `undefined`.

Found this while trying to bump to `v11.20.0`.

This was introduced in:
* https://github.com/Shopify/polaris/pull/10566