[polaris.shopify.com] Fix layout-tokens do usages (#10951)
### WHY are these changes introduced?

Due to the way that DirectiveCard has been
[written](https://github.com/Shopify/polaris/pull/10597), whitespace is
necessary in order for the mdx compile to consider the image as a
separate node. Otherwise it will wrap the image and the paragraph below
it in a single p tag.

### WHAT is this pull request doing?

Just adds white space between the image tag and the paragraph below.