Replace path fill with svg fill (#11262)
### WHY are these changes introduced?

`fill="currentColor"` is sometimes repeated 3-4 times depending on the
number of `<path>` elements in an `<svg>`. This PR moves
`fill="currentColor"` to the root `<svg>` element to simplify how the
color flows through the icon. The most complicated icons like
`SortAscending` and `SortDescending` which use `fill-opacity` still work
with the `fill` value moved to the parent. I also searched for `circle`
and `polygon` elements but they were unused.

### WHAT is this pull request doing?

Reducing the size of our SVG icons by moving the `fill` property to the
`<svg>`