[IndexTable] Only run sticky logic if we have a defined rootBounds object  (#11584)
### WHY are these changes introduced?

Sometimes when rendering IndexTables within iframes, the `rootBounds`
object on the `IntersectionObserverEntry` entity can be `null`, rather
than an object containing dimensions of the root itself. We were
defaulting the `entry.rootBounds.height` value to 0 if it didn't exist.
This was causing these IndexTables to incorrectly calculate whether the
SelectAllActions was sticky or not. We should not change the sticky
state of the SelectAllActions if we have no `rootBounds` object as we
wouldn't be able to accurately determine the sticky state.

### How to 🎩

Codesandbox link:
https://codesandbox.io/p/sandbox/lucid-cherry-pc5tyr?file=%2Fpackage.json%3A5%2C23

Try changing the Polaris version to `latest` in `package.json` and
toggle the IndexTable a few times to see the issue happening. Then
confirm that with the snapshot it no longer happens.

### 🎩 checklist

- [x] Tested a
[snapshot](https://github.com/Shopify/polaris/blob/main/documentation/Releasing.md#-snapshot-releases)
- [x] Tested on
[mobile](https://github.com/Shopify/polaris/blob/main/documentation/Tophatting.md#cross-browser-testing)
- [x] Tested on [multiple
browsers](https://help.shopify.com/en/manual/shopify-admin/supported-browsers)
- [x] Tested for
[accessibility](https://github.com/Shopify/polaris/blob/main/documentation/Accessibility%20testing.md)
- [x] Updated the component's `README.md` with documentation changes
- [x] [Tophatted
documentation](https://github.com/Shopify/polaris/blob/main/documentation/Tophatting%20documentation.md)
changes in the style guide

---------

Co-authored-by: Chloe Rice <chloerice@users.noreply.github.com>