Create helper function and update all Combobox examples
As part of the filtering process, the example code creates a RegExp out
of the user's input. For all characters in the input that have a special
meaning in regular expressions, we must replace them with a \ followed
by that character, or else an error will be thrown if a user enters any
"special" regular expression characters.

Now RegExp interprets the user's input just as text.