[Actions] Ensure incorrect values are filtered out before iterating (#11684)
### WHY are these changes introduced?

Fixes a
[bug](https://app.bugsnag.com/shopify/shopify-web-client/timeline?filters[event.since]=7d&filters[error.status]=open&filters[error]=65e257c1fa853900080d195d&filters[release.seen_in]=172541221c00bcb5afbd4779e553710d0eaed216&pivot_tab=event)
that was detected last week where incorrect value types were passed down
to the `Actions` component within the `Page` component which was causing
pages to 500.

### WHAT is this pull request doing?

We create an array of hidden actions and hidden groups to render within
the ActionList, depending on how much space there is to render regular
actions & groups. We call a typeguard function on each item in that
array, using the `'X' in y` syntax to determine the type of the value.
Some of the values in the array when used in production are `undefined`,
so trying to run `'X' in undefined` throws an error. By filtering the
data that creates the array which we iterate on, we can be sure that the
values exist before we pass them to the typeguard function.

### How to 🎩

🖥 [Local development
instructions](https://github.com/Shopify/polaris/blob/main/README.md#install-dependencies-and-build-workspaces)
🗒 [General tophatting
guidelines](https://github.com/Shopify/polaris/blob/main/documentation/Tophatting.md)
📄 [Changelog
guidelines](https://github.com/Shopify/polaris/blob/main/.github/CONTRIBUTING.md#changelog)

Spin instance:
https://admin.web.more-actions-bug-2.marc-thomas.eu.spin.dev/store/shop1/products/10

Should not be able to get an error in the console relating to the bug if
you navigate between products with long and short titles

### 🎩 checklist

- [x] Tested a
[snapshot](https://github.com/Shopify/polaris/blob/main/documentation/Releasing.md#-snapshot-releases)
- [x] Tested on
[mobile](https://github.com/Shopify/polaris/blob/main/documentation/Tophatting.md#cross-browser-testing)
- [x] Tested on [multiple
browsers](https://help.shopify.com/en/manual/shopify-admin/supported-browsers)
- [x] Tested for
[accessibility](https://github.com/Shopify/polaris/blob/main/documentation/Accessibility%20testing.md)
- [x] Updated the component's `README.md` with documentation changes
- [x] [Tophatted
documentation](https://github.com/Shopify/polaris/blob/main/documentation/Tophatting%20documentation.md)
changes in the style guide