Fix regression with sidebar styling
A previous [PR](https://github.com/Shopify/polaris/pull/11628)
attempting to fix an unwanted border next to the sidebar unintentionally
caused a regression in Admin: app pages in Admin no longer maintain
space on the right for the universal sidebar.

The fix: revert the change from the previous PR which changed margin to
padding. Use margin, but instead of rendering `border-inline-end` on
`.Content`, which causes the unwanted border, render it on `.Main`
instead.