Pin GitHub Actions to commit (#13569)
This is an automated PR to update actions in this repo. The operation
should be no-op, as we are only switching out the version tag with the
matching commit SHA.

To align with industry best practices, we are going to pin Github
Actions to a specific commit SHA.

To read more about why pinning actions is recommended check
[here](https://docs.github.com/en/actions/security-guides/security-hardening-for-github-actions#using-third-party-actions).

To ensure these Actions stay to-do-date, this PR also enables Dependabot
automated updates. To read more about this configuration check
[here](https://docs.github.com/en/code-security/dependabot/working-with-dependabot/keeping-your-actions-up-to-date-with-dependabot).

If allowed, this PR will attempt to auto-merge in ~1 week.